Implement fix for hydroponics tray allowing autogrow while no power (#86723)
## About The Pull Request

Fixes #83539.

The code was structured in such a way that the check for power usage did
not work correctly, and this allowed for someone to double press ctrl
click on the hydroponics tray while there is no power, turning it on
forever regardless.
## Why It's Good For The Game

Fixes a bug that allows autogrow to be turned on with no power.
## Changelog
:cl:
fix: fixed a bug that allows autogrow to be turned on with no power.
/:cl: