Implemenet an alert that blob expansion failed in space (#86728)
## About The Pull Request

Adds a balloon alert that the expansion has failed, as there previously
was confusion on why the blob was attacking space. See #84618 for such
an instance.
## Why It's Good For The Game

Makes it clear that the blob is not just attacking space for no reason,
it is just the expansion failing.
## Changelog
:cl:
qol: made it clearer that the expansion in space failed for blob instead
of wildly flailing at space.
/:cl: