Security Flashbangs cannot be instantly detonated anymore. (#86704)
## About The Pull Request

Removes the option of setting a sec flashbang to instant detonation.

## Why It's Good For The Game

My reasoning from my previous PR hasn't really changed, so excuse me for
the few copypasted bits.

Flashbang grenades are one of the best tools in Security arsenal. they
are available round-start, widely abundant, and part of the reason why
they are so strong is how rare ear protection tends to be.

But what tilts the scale from strong to just unfair, in my opinion, is
the combination of instant detonation and being able to pierce ear
protection entirely on adjacent tiles.

This is a hardcoded stun by the way, You could literally have no eyes
nor ears and it won't make a difference.

I initially considered just reducing the radius at which penetration
occurs, but i failed to consider how it would probably result in an
unintended buff due to the ability to drag a primed flashbang.

So instead, after discussing options with a few other people, including
mantainers I think the sanest option is to remove the ability for
instant detonation.

You can still technically run at someone as a desperate attempt to
disarm them, but they'll have a sliver of chance to react.

Flashbangs should be used as grenades, not as instant unreactable
hardstuns, this strategy requiring teamwork to pull off doesn't justify
it being completely unfair and unreactable on the receiver.

## Changelog

:cl:
balance: Security Flashbangs can no longer be primed for instant
detonation.
/:cl:

---------

Co-authored-by: Xander3359 <66163761+Xander3359@users.noreply.github.com>
Co-authored-by: Time-Green <7501474+Time-Green@users.noreply.github.com>