Various syndicate crossbow adjustments (#86781)
## About The Pull Request
Adjustments have been made to the Syndicate Crossbow to address its
current strength and balance issues. The weapon is being heavily abused
due to its overpowering combination of damage and utility. These changes
aim to reduce its effectiveness while maintaining its role in the game.

List of Changes:
    Syndicate Rebar Bolt Damage: Reduced to 45 from 55.
    Hydrogen Bolt Damage: Reduced 35 from 55.
    Hydrogen Bolt Wound chance: reduced to 0 to prevent delimbing.
    Quiver Reload: Reload is now interrupted by movement.
    Quiver Reload: Reload increased 1.2 seconds from 0.8 seconds.
    Crossbow TC Cost Increased to 12 from 10.
    Quiver size: Increased to normal from small.

Edit : after some more thought and looking at the original creator's
ideal, i've made some more changes. I've adjusted damage slightly to
still fit in damage breakpoints but not rely on mood. Syndicate crossbow
was meant to be a sidegrade to the syndicate revolver, it is currently
an upgrade, even with these changes.

## Why It's Good For The Game
The Syndicate Crossbow’s high damage output, especially with hydrogen
bolts, makes it overly powerful, leading to imbalanced gameplay.
Reducing the damage brings it more in line with other weapons like the
syndicate revolver while ensuring it still remains useful in certain
situations.

We should consider this weapon as a sidegrade to the syndicate revolver,
we should look at what armor does to the damage. With these changes, we
kill both a security officer and an unarmored crew member in 3 shots.
The revolver kills an unarmored crew member in 2 shots but an armored
security officer in 3 or 4, depending on their mood.

Comparing the Syndicate crossbow to the Syndicate Revolver, it is easier
to find ammo for, has higher armor pen, is easier to store ammo for,
easier to reload, has a scope, and the ammo is reusable. The crossbow
needs a bit more of a trade-off than some reduced ammo capacity.

Metal hydrogen bolts fill a niche for silently killing pets across the
map, not killing the entire station from a 1x1 hole in maintenance. They
have no reason to be doing as much damage as syndicate rebar bolt damage
or close to zaukerite bolts which are significantly more difficult to
make.

With the introduction of the syndicate quiver, the syndicate crossbow
has gotten significantly more powerful. Much of the clunkiness that made
it 'worse' than the syndicate revolver has been removed (similar TC
cost, damage, utility. 13 TC for revolver, 10 TC for crossbow). The
increase in power deserves an increase in TC cost. The syndicate
crossbow was meant to be a sidegrade to the syndicate revolver, we can
adjust the price so balance between them is more cut and dry.

The ability to reload the crossbow with the quiver while moving is also
a bit too strong. It effectively gives you a magazine of 20 shots while
in active combat. You can spam your reload keybind and left mouse button
while moving to continue firing at a rate of one shot per 0.8 seconds. I
feel like this kind of defeats the entire idea behind the crossbow.
Edit : I found while testing that I could still reload while moving at
0.8 second reload if i stutter step, I could still reload during combat.
I've increased it to 1.2 seconds.

Quiver size was increased because it's silly for a quiver with 20 bolts
to be able to fit in a box

## Changelog
:cl:
balance: Syndicate Rebar Bolt damage reduced to 45 from 55.
balance: Hydrogen Bolt damage reduced to 35 from 55.
balance: Syndicate Quiver reload is now interrupted by movement.
balance: Syndicate Quiver reload increased to 1.2 seconds from 0.8
seconds.
balance: Crossbow TC cost increased to 12 from 10.
balance: Quiver size increased to normal from small.
/:cl: