Fix the greeting message you get upon becoming an imaginary friend including nulls (#86866)
## About The Pull Request

Previously the imaginary friend greeting message would be called on
`Login()`, but this meant that it could run before it actually linked
the `owner` value, thus putting nulls in the message where the owner's
name should've been.
This just changes it so `greet()` is only called on `Login()` if there's
already an owner linked, otherwise calling it when being attached to an
owner.
## Why It's Good For The Game

Better if the greeting message actually says whose imaginary friend you
are.
## Changelog
:cl:
fix: The greeting message imaginary friends get upon becoming one
actually includes the owner's name, instead of displaying nothing where
it should've been.
/:cl: