Fixes water overlays showing on toilets facing directions other than SOUTH (#86878)
## About The Pull Request
The overlays shows on directions other than south because the code
doesn't update the overlays whenever the dir is changed.
Honestly saving a tiny blip of space with a pair of 1-dir static icons
isn't worth making all this messy code over anyway, so I just scratched
all of it and gave the icon state 4 dirs (3 of which are empty but that
may change if someone decides to give the toilet
a 3/4 perspective when facing right or left.

## Why It's Good For The Game
Fixed water overlays showing on toilets facing directions other than
SOUTH (down).

## Changelog

:cl:
fix: Fixed water overlays showing on toilets facing directions other
than SOUTH (down).
/:cl: