Removes intentional ling check in health analyzer (#86893)
## About The Pull Request

Advanced health analyzers have an intended ling check which tells you if
the scanned body is faking death (is in regen stasis), which is at odds
with the rest of the codebase (which actively removes unintentional ling
checks and blocks the addition of intentional ones).

This PR removes it (making advanced analyzers unable to determine if the
target is faking death).
Also tweaks some code around it in general.

## Why It's Good For The Game

I'm actually like 50/50 on whether this should actually be removed or
not, so I opened the PR for discussion.

On one hand I think it's kind of interesting to have a doctor look at
the results *closely* and realize something is wrong here.
On the other hand, ling check bad?

## Changelog

:cl: Melbert
del: Advanced Health Analyzers can no longer be used to determine if a
body is faking death (like a changeling)
/:cl: