Removes a bookcase from the holdout simulation (#86907)
## About The Pull Request
Gets rid of a bookcase that was under the foam in the holdout bunker
simulation.
## Why It's Good For The Game
There's no reason for it to be there, and it would cause random books to
be there after simulations changed. And most importantly, it caused me
MILD ANNOYANCE.
## Changelog
:cl: Goat
fix: Removes a random bookcase that was in the holdout bunker holodeck
simulation
/:cl: